Modular structure of celery client
